+++ src/model.rs
@@ -1,6 +1,8 @@
+use rust_decimal::Decimal;
+
 #[derive(Debug, serde::Deserialize, serde::Serialize, Clone, PartialEq)]
 pub struct ADX {
-    pub adx_opt: Option<f64>,
-    pub di_plus_opt: Option<f64>,
-    pub di_minus_opt: Option<f64>,
+    pub adx_opt: Option<Decimal>,
+    pub di_plus_opt: Option<Decimal>,
+    pub di_minus_opt: Option<Decimal>,
 }
